container operations divided into separate files
